Improved userfow to include refresh data trigger
added a refresh button using streamlit. when clicked, it scrapes all the data, chunks it, creates embeddings, creates a faiss index and stores it. until this task is complete, disable the query text inpt and the search button.

similarly, if the user has already clicked on search, disable the data refresh button until the search results are received